fix(FEV-1255): displays timeline preview only for video type of entries
